Title: 괄호 쌍 만들어주기 3, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 231209/괄호 쌍 만들어주기 3/pair-parentheses-3.cpp
@@ -6,8 +6,8 @@ string s;
 int ans;
 int main(){
     cin >> s;
-    for(int i=0; i<s.size(); i++){
-        for(int j=i+1; j<s.size(); j++){
+    for(int i=0; i<(int)s.size(); i++){
+        for(int j=i+1; j<(int)s.size(); j++){
             if(s[i] == '(' && s[j] == ')'){
                 ans++;
             }
